requirements.txt
absl-py==2.2.2
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
braceexpand==0.1.7
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.2
cycler==0.12.1
Cython==3.1.1
cytoolz==1.0.1
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
Distance==0.1.3
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fiddle==0.3.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.12.0
future==1.0.0
g2p-en==2.1.0
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
inflect==7.5.0
iniconfig==2.1.0
intervaltree==3.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.5.1
jwcrypto==1.5.6
kaldi-python-io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.30.3
libcst==1.7.0
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.44.0
loguru==0.7.3
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
nemo-toolkit==2.3.0
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
oauth2client==4.1.3
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.22.0
optuna==4.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.2
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.24.4
psutil==7.0.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyloudnorm==0.1.1
pyparsing==3.2.3
pyreadline3==3.5.4
Pyrebase4==4.8.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jwt==4.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==0.10.1
resampy==0.4.3
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==0.5.0.post1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.6.0
tokenizers==0.21.1
toolz==1.0.0
torch==2.7.0
torchmetrics==1.7.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
typeguard==4.4.2
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.2
wandb==0.19.11
watchfiles==1.0.5
wcwidth==0.2.13
webdataset==0.2.111
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
win32_setctime==1.2.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
